+++ crates/cli/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## 0.4.1-git
+## 0.5.0-git
+
+### Major
+
+- Change `platform-update` to take a bundle instead of the 2 sides
 
 ### Patch
 
@@ -97,4 +101,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 20 -->
+<!-- Increment to skip CHANGELOG.md test: 21 -->